fix: build error

+++ include/util/tcompression.h
@@ -54,7 +54,7 @@ extern "C" {
 #ifdef TD_TSZ
 extern bool lossyFloat;
 extern bool lossyDouble;
-int32_t tsCompressInit(char* lossyColumns, double fPrecision, double dPrecision, uint32_t maxIntervals, uint32_t intervals,
+int32_t tsCompressInit(char* lossyColumns, float fPrecision, double dPrecision, uint32_t maxIntervals, uint32_t intervals,
                        int32_t ifAdtFse, const char* compressor);
 
 void    tsCompressExit();
